(WIP-ish) custom presets implementation
In case a future board maker decides to create some kind of hat for any particular board, and feels as if the default layout is not appropriate for their uses, the new preset layouts combobox in the board layout tab can be used to select any one of these, if the given board has them.

Right now, all it provides is a reference (i.e. test) entry for the rpipico family, and working presets for both SAMCO 1.1 and 2.0 boards.