Temperature threshold updates
Sync temperature stuff with the new expected values - the old ones were too damn high. Also adds "°C" to clarify the (estimated) units of measurement.
Since the values are planned to be configurable in a later update and should be pulled from settingsTable when they are, these temperature values are just the current defaults.